Add explicit cast to shut up gcc -W.

+++ src/config.c
@@ -200,7 +200,7 @@ int inet_aton(const char *cp, struct in_addr *pin)
 	unsigned long ia;
 
 	ia = inet_addr(cp);
-	if (ia == -1)
+	if (ia == (unsigned long) -1)
 		return 0;
 	pin->s_addr = ia;
 	return 1;
